# Conflicts:
#	.idea/misc.xml
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/id/thork/app/repository/TaskRepository.kt

+++ app/src/main/java/id/thork/app/pages/material_plan/element/MaterialPlanViewModel.kt
@@ -33,7 +33,7 @@ class MaterialPlanViewModel @ViewModelInject constructor(
 
 
     fun initListMaterialPlan(workorderid: String) {
-        val materialPlan = materialRepository.getListMaterialPlanByWoid(workorderid)
+        val materialPlan = materialRepository.getListMaterialPlanByWoid(workorderid.toInt())
         materialPlan.whatIfNotNull {
             _listMaterial.value = it
         }
